# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COMFACA Crédito API",
        "description": "API para la gestión de solicitudes de crédito de COMFACA",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Obtener notificaciones del usuario",
                "description": "Obtener todas las notificaciones del usuario autenticado\n\nGET /api/notifications\n\nQuery params:\n- unread: boolean (opcional) - Solo notificaciones no leídas\n- limit: int (opcional) - Límite de resultados (default: 50)",
                "operationId": "b1274a2aaae8a4295a7e1925e8d8090f",
                "parameters": [
                    {
                        "name": "unread",
                        "in": "query",
                        "description": "Solo no leídas",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Límite de resultados",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificaciones obtenidas"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/unread-count": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Obtener contador de notificaciones no leídas",
                "description": "Obtener contador de notificaciones no leídas\n\nGET /api/notifications/unread-count",
                "operationId": "85697ee523a5dace428185c95efb9f2f",
                "responses": {
                    "200": {
                        "description": "Contador obtenido"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}/read": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marcar notificación como leída",
                "description": "Marcar una notificación como leída\n\nPUT /api/notifications/{id}/read",
                "operationId": "8bc5203b95eaa5cf8614a66ced103386",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la notificación",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación marcada como leída"
                    },
                    "404": {
                        "description": "Notificación no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/mark-all-read": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marcar todas las notificaciones como leídas",
                "description": "Marcar todas las notificaciones como leídas\n\nPUT /api/notifications/mark-all-read",
                "operationId": "48969f1c1523374341cf86b5610b3e03",
                "responses": {
                    "200": {
                        "description": "Notificaciones marcadas como leídas"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}": {
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Eliminar notificación",
                "description": "Eliminar una notificación\n\nDELETE /api/notifications/{id}",
                "operationId": "0f28a0400610f81f01e2d25ce4e0edff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la notificación",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación eliminada"
                    },
                    "404": {
                        "description": "Notificación no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}